refactor: getStatusBackgroundColor をui.commonに移動させた
